+++ src/functions/scalar/fusion_relative/implementation.cpp
@@ -3,7 +3,7 @@
 namespace flockmtl {
 
 void FusionRelative::ValidateArguments(duckdb::DataChunk& args) {
-    for (int i = 0; i < args.ColumnCount(); i++) {
+    for (int i = 0; i < static_cast<int>(args.ColumnCount()); i++) {
         if (args.data[i].GetType() != duckdb::LogicalType::DOUBLE) {
             throw std::runtime_error("fusion_relative: argument must be a double");
         }
@@ -14,9 +14,9 @@ std::vector<double> FusionRelative::Operation(duckdb::DataChunk& args) {
     FusionRelative::ValidateArguments(args);
 
     std::vector<double> results;
-    for (auto i = 0; i < args.size(); i++) {
+    for (auto i = 0; i < static_cast<int>(args.size()); i++) {
         auto max = 0.0;
-        for (auto j = 0; j < args.ColumnCount(); j++) {
+        for (auto j = 0; j < static_cast<int>(args.ColumnCount()); j++) {
             auto valueWrapper = args.data[j].GetValue(i);
             if (!valueWrapper.IsNull()) {
                 auto value = valueWrapper.GetValue<double>();
